Ebenezer support for parquet with unions
Due to a scrooge bug (until release 3.19), parquet formats containing unions
could not be supported in ebenezer (metadata for structs would make the
execution abort with NoSuchMethodException).

This patch introduces a workaround: metadata are retrieved in a stand-alone
helper (specific to scrooge 3.17), and for union member fields, the exception
is caught and the metadata are processed with this in mind.

The 3.19 version of scrooge offers special metadata for union fields, but
going to 3.19 would entail many changes in the dependencies of projects,
and substantially different changes to the code changed here.