Changes `Hive.createDatabase` and `createTable` to check existance upfront.
Changes the Hive create functions to check if the database and/or table
already exists before trying to create them. While Hive handles trying
to create an already existing table/database correctly it causes
exceptions to be logged to the output. This doesn't get rid of the
problem that Hive logs exceptions that are being handled instead it
prevents some of the exceptions by checking before creating a database/table.

It also adds the `existsDatabase` method to `Hive`.